[fix]補上內頁登入保護與資料存取權限控制

+++ MyNote/src/main/java/com/mynote/service/NoteService.java
@@ -2,7 +2,10 @@ package com.mynote.service;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.mynote.mapper.NoteMapper;
 import com.mynote.model.NoteModel;
@@ -10,11 +13,8 @@ import com.mynote.model.NoteModel;
 @Service
 public class NoteService {
 
-    private final NoteMapper noteMapper;
-
-    public NoteService(NoteMapper noteMapper) {
-        this.noteMapper = noteMapper;
-    }
+    @Autowired
+    private NoteMapper noteMapper;
 
     // 依"使用者-userId"取得筆記
     public List<NoteModel> findByUserId(Integer userId) {
@@ -29,14 +29,31 @@ public class NoteService {
     }
 
     // 更新筆記
-    public NoteModel update(NoteModel note) {
+    public NoteModel update(Integer userId, NoteModel note) {
+        NoteModel existingNote = noteMapper.findById(note.getId());
+        validateOwnership(userId, existingNote);
+
+        note.setUserId(existingNote.getUserId());
+        note.setCreatedAt(existingNote.getCreatedAt());
         noteMapper.update(note);
-        return note;
+        return noteMapper.findById(note.getId());
     }
 
     // 刪除筆記
-    public void delete(Integer id) {
+    public void delete(Integer userId, Integer id) {
+        NoteModel existingNote = noteMapper.findById(id);
+        validateOwnership(userId, existingNote);
+
         System.out.println("Deleting note ID: " + id);
         noteMapper.deleteById(id);
+    }
+
+    private void validateOwnership(Integer userId, NoteModel note) {
+        if (note == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Note not found.");
+        }
+        if (!userId.equals(note.getUserId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You cannot access this note.");
+        }
     }
 }
